third_party/WebKit/LayoutTests/bluetooth/characteristic/readValue/characteristic-is-removed.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../../../external/wpt/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
const test_desc = 'Characteristic gets removed. Reject with InvalidStateError.';
let characteristic, fake_characteristic;

bluetooth_test(() => getMeasurementIntervalCharacteristic()
    .then(_ => ({characteristic, fake_characteristic} = _))
    .then(() => fake_characteristic.remove())
    .then(() => assert_promise_rejects_with_message(
      characteristic.readValue(),
      new DOMException(
        'GATT Characteristic no longer exists.',
        'InvalidStateError'),
      'Characteristic got removed.')),
    test_desc);
</script>
